ADR-002 stage 2.6 (PR2/3): split agent-dispatch intents out of register_canvas
Relocates the specialized agent-dispatch node kinds - Web Research,
Artifact, Chart, the note-agent/branch-comparison group, Gitlink,
Py-Coder, and Execution Sandbox - out of register_canvas's ~1570-line
body into seven feature-scoped backend/api/intents_*.py modules,
continuing the pattern PR1 established for the core chat/node surface.
Pure code motion: every closure is relocated verbatim, byte-diffed
against the pre-PR2 baseline to confirm no logic changed, with only two
intentional comment updates where a "see X above" reference needed to
name its new target module instead.

register_canvas is now a thin orchestrator that constructs the shared
SceneDocument and calls each register_*_intents function in turn; the
remaining notes/frames/groups/view/grid intents are PR3's scope.
