+++ backend/canvas.py
@@ -21,19 +21,13 @@ stay here, as does the entire register_canvas closure set.
 
 from __future__ import annotations
 
-import os
 from typing import Any
 
-from graphlink_chart_data import ChartDataError, canonicalize_chart_data, SUPPORTED_CHART_TYPES
+from graphlink_chart_data import SUPPORTED_CHART_TYPES
 from graphlink_grid_view_settings import GRID_STYLE_PRESETS
 from graphlink_navigation_pins import NavigationPinRecord
 
-from backend import native_dialogs
-from backend.agents import (
-    _CODE_EXEC_RUN_CLAIM_PLACEHOLDER,
-    _GITLINK_RUN_CLAIM_PLACEHOLDER,
-    AgentDispatcher,
-)
+from backend.agents import AgentDispatcher
 from backend.composer import ComposerDocument
 from backend.events import SessionBus
 from backend.notifications import NotificationState
@@ -239,10 +233,17 @@ def _placeholder_chart_data(chart_type: str) -> dict[str, Any]:
 # namespace, so intents_chat.py's own `from backend.canvas import
 # _history_token_text` (evaluated when the import below first pulls that
 # module in) succeeds instead of hitting a partially-initialized module.
+from backend.api.intents_artifact import register_artifact_intents  # noqa: E402
+from backend.api.intents_branches import register_branches_intents  # noqa: E402
+from backend.api.intents_chart import register_chart_intents  # noqa: E402
 from backend.api.intents_chat import register_chat_intents  # noqa: E402
 from backend.api.intents_chat_image import register_chat_image_intents  # noqa: E402
+from backend.api.intents_code_sandbox import register_code_sandbox_intents  # noqa: E402
 from backend.api.intents_conversation import register_conversation_intents  # noqa: E402
+from backend.api.intents_gitlink import register_gitlink_intents  # noqa: E402
 from backend.api.intents_nodes import register_node_intents  # noqa: E402
+from backend.api.intents_pycoder import register_pycoder_intents  # noqa: E402
+from backend.api.intents_web_research import register_web_research_intents  # noqa: E402
 
 
 def register_canvas(
@@ -313,718 +314,13 @@ def register_canvas(
     register_chat_intents(bus, document, notifications, agent_dispatcher, composer_document, token_counter)
     register_chat_image_intents(bus, document, notifications, agent_dispatcher)
 
-    async def run_web_research(node_id, query_text):
-        if agent_dispatcher.is_web_research_busy():
-            # Checked BEFORE touching document state: start_web_research_run
-            # resets a node's progress/error fields unconditionally, and the
-            # dispatcher only allows one web-research run at a time anyway -
-            # without this early check, clicking Run on a different node
-            # while one is already in flight would silently wipe that node's
-            # prior result/error banner even though no new run actually starts.
-            notifications.show("A web research request is already running.", "info")
-            await bus.publish("notification")
-            return None
-        try:
-            node = document.start_web_research_run(node_id, query_text)
-        except SceneError:
-            notifications.show("This node no longer exists.", "warning")
-            await bus.publish("notification")
-            return None
-        await publish_scene()
-
-        parent_edge = document._branch_parent_edge(node_id)
-        branch_history = document.chat_branch_history(parent_edge.source) if parent_edge else []
-
-        async def _on_progress(event):
-            if node_id not in document.nodes:
-                return
-            document.apply_web_research_progress(node_id, event)
-            await bus.publish("scene")
-
-        async def _on_success(result):
-            if node_id not in document.nodes:
-                return
-            document.complete_web_research_run(node_id, _research_result_wire(result))
-            await bus.publish("scene")
-
-        async def _on_failure(exc):
-            if node_id not in document.nodes:
-                return
-            cancelled = type(exc).__name__ == "RequestCancelled"
-            document.fail_web_research_run(node_id, cancelled=cancelled, message=str(exc))
-            await bus.publish("scene")
-
-        await agent_dispatcher.start_web_research(
-            bus=bus,
-            notifications_state=notifications,
-            node=node,
-            node_id=node_id,
-            query=query_text,
-            branch_history=branch_history,
-            on_progress=_on_progress,
-            on_success=_on_success,
-            on_failure=_on_failure,
-        )
-        return node_id
-
-    async def cancel_web_research_request(request_id):
-        agent_dispatcher.cancel_web_research(request_id)
-
-    async def send_artifact_message(node_id, text):
-        # R5.2: the Artifact node's own Send action - appends a real user
-        # instruction, then dispatches a real agent reply through
-        # AgentDispatcher.start_artifact_reply. No try/except SceneError guard
-        # here (an unknown node_id propagates as a generic WS intent error) -
-        # same posture as send_conversation_message (backend/api/
-        # intents_conversation.py, ADR-002 stage 2.6), not
-        # run_web_research's defensive pre-check pattern above: there is no
-        # persisted progress/error state on this node that an unguarded call
-        # could corrupt, so a stale click racing a delete has nothing
-        # destructive to protect against.
-        node = document.send_artifact_message(node_id, text)
-        await publish_scene()
-
-        parent_edge = document._branch_parent_edge(node_id)
-        branch_history = document.chat_branch_history(parent_edge.source) if parent_edge else []
-        full_history = branch_history + node.history
-
-        def _on_reply(new_content, ai_message):
-            document.complete_artifact_generation(node_id, new_content, ai_message)
-
-        await agent_dispatcher.start_artifact_reply(
-            bus=bus,
-            notifications_state=notifications,
-            node=node,
-            current_artifact=node.artifact_content,
-            history=full_history,
-            on_reply=_on_reply,
-        )
-        return node.id
-
-    async def cancel_artifact_request(request_id):
-        agent_dispatcher.cancel_artifact(request_id)
-
-    # -- R6.2: Chart node ------------------------------------------------------
-    #
-    # Unlike every branch-point-child kind above (Web Research/Artifact/
-    # Gitlink/Py-Coder/Execution Sandbox), Chart has no separate "create an
-    # empty node, then run generation on it" split - generateChart is a
-    # single combined create+generate action, so there is no addChartNode
-    # intent at all: the SceneNode is only ever created (by
-    # document.add_chart_node, in _on_success below) once real chart data
-    # actually exists, mirroring legacy's own ChartWorkerThread flow (a
-    # transient loading state anchored on the SOURCE node, not a pre-created
-    # placeholder chart node - see graphlink_window_actions.py's
-    # generate_chart/handle_chart_data).
-
-    async def generate_chart(parent_node_id, chart_type):
-        if not parent_node_id or parent_node_id not in document.nodes:
-            notifications.show(
-                "Please select a valid node to branch from before generating a chart.",
-                "warning",
-            )
-            await bus.publish("notification")
-            return None
-
-        normalized_chart_type = str(chart_type or "").strip().lower()
-        if normalized_chart_type not in SUPPORTED_CHART_TYPES:
-            notifications.show(
-                "Please choose a valid chart type before generating a chart.",
-                "warning",
-            )
-            await bus.publish("notification")
-            return None
-
-        parent = document.nodes[parent_node_id]
-        branch_history = document.chat_branch_history(parent_node_id)
-        source_text = _chart_source_text(branch_history)
-
-        result_holder: dict[str, str] = {}
-
-        async def _on_success(result):
-            try:
-                chart_data = canonicalize_chart_data(result, normalized_chart_type)
-                chart_error = ""
-            except ChartDataError as exc:
-                # R6.2 contract: ChartDataAgent's own validate_chart_data
-                # pipeline (repair round trip, then heuristic fallback)
-                # already tries hard to guarantee canonical output before
-                # ever returning successfully - this is the rare defensive
-                # case where it still somehow didn't. Never a silent no-op:
-                # still create a real chart node with a minimal placeholder
-                # shape and chart_error set, same "degrade gracefully, never
-                # drop the request" contract as the agent's own internal
-                # fallback chain.
-                chart_data = _placeholder_chart_data(normalized_chart_type)
-                chart_error = f"The generated chart data could not be validated: {exc}"
-            node = document.add_chart_node(
-                parent.x + MESSAGE_VERTICAL_SPACING,
-                parent.y,
-                parent_node_id,
-                normalized_chart_type,
-                chart_data,
-                chart_error=chart_error,
-            )
-            result_holder["node_id"] = node.id
-            await bus.publish("scene")
-
-        def _on_failure(message):
-            # Matches ChartWorkerThread's own error path (and this feature's
-            # explicit contract): a genuinely unrecoverable agent-side
-            # failure (get_response's own top-level "error" key, or a
-            # timeout/exception) shows a notification and creates nothing -
-            # start_chart_generation above already shows that notification,
-            # so there is nothing left for this callback to do.
-            pass
-
-        await agent_dispatcher.start_chart_generation(
-            bus=bus,
-            notifications_state=notifications,
-            node_id=parent_node_id,
-            chart_type=normalized_chart_type,
-            source_text=source_text,
-            on_success=_on_success,
-            on_failure=_on_failure,
-        )
-        return result_holder.get("node_id")
-
-    async def _generate_note_from_node(source_node_id, note_kind, x_offset, y_offset):
-        """R8a: shared path for generateKeyTakeaway and generateExplainerNote.
-
-        Both take one chat node, run its text through an agent, and drop the
-        result into a new note beside it - identical except for the agent and
-        the note's offset, so they share one implementation rather than two
-        that can drift.
-
-        Source text is the node's OWN content, not the branch history that
-        generate_chart uses: legacy's takeaway/explainer passed a single
-        node's text, and widening that to the whole branch would change what
-        the feature summarises.
-        """
-        if not source_node_id or source_node_id not in document.nodes:
-            notifications.show("Please select a valid node first.", "warning")
-            await bus.publish("notification")
-            return None
-
-        source = document.nodes[source_node_id]
-        if source.kind != "chat":
-            notifications.show("This node can't be summarised into a note.", "warning")
-            await bus.publish("notification")
-            return None
-        if not source.content or not source.content.strip():
-            notifications.show("The selected node has no text to summarise.", "warning")
-            await bus.publish("notification")
-            return None
-
-        result_holder: dict[str, str] = {}
-
-        async def _on_success(text):
-            if source_node_id not in document.nodes:
-                # Deleted mid-flight - silent no-op, same posture as
-                # _dispatch_image's own liveness check.
-                return
-            note = document.add_note(source.x + x_offset, source.y + y_offset)
-            document.set_note_content(note.id, text)
-            # Legacy tinted these notes "Mid Gray" with an info-coloured
-            # header. Both values come from the frontend's own palette
-            # (GroupColorPicker's GROUP_MONO_COLORS/GROUP_NAMED_COLORS) since
-            # the backend stores hex and never resolves a colour name. The
-            # legacy note width of 400 is NOT ported: note width is not a
-            # modeled field here (it is CSS-driven), so there is nothing to
-            # set it on.
-            document.set_group_color(note.id, NOTE_AGENT_BODY_COLOR, NOTE_AGENT_HEADER_COLOR)
-            result_holder["node_id"] = note.id
-            await bus.publish("scene")
-
-        def _on_failure(message):
-            # start_note_generation already surfaced the notification.
-            pass
-
-        await agent_dispatcher.start_note_generation(
-            bus=bus,
-            notifications_state=notifications,
-            node_id=source_node_id,
-            note_kind=note_kind,
-            source_text=source.content,
-            on_success=_on_success,
-            on_failure=_on_failure,
-        )
-        return result_holder.get("node_id")
-
-    async def generate_key_takeaway(source_node_id):
-        return await _generate_note_from_node(source_node_id, "takeaway", NOTE_AGENT_X_OFFSET, 0)
-
-    async def generate_explainer_note(source_node_id):
-        # Offset vertically as well as horizontally so a takeaway and an
-        # explainer generated from the same node don't land on top of each
-        # other - the same 100px stagger legacy used.
-        return await _generate_note_from_node(source_node_id, "explainer", NOTE_AGENT_X_OFFSET, 100)
-
-    async def compare_branches(node_ids):
-        """ADR-002 Workstream 1 ("Compare Branches") - the second sequenced
-        item after "Branch from here" (that same workstream's fork
-        primitive). Takes 2+ existing chat nodes, walks each one's own
-        chat_branch_history, and drops a single agent-authored comparison
-        into a new note linked back to every source branch (note.item_ids
-        - see mark_branch_comparison_note).
-
-        Deliberately no auto-selection fallback, unlike the single-node
-        note agents above: there is no sensible single "the selected node"
-        default here - the caller (App.tsx's own Compare Branches shortcut)
-        must supply 2+ real ids up front, the same "the frontend already
-        gathered React Flow's own multi-selection" contract create_frame/
-        create_container already use."""
-        ids = list(dict.fromkeys(str(i) for i in (node_ids or [])))  # de-dupe, preserve order
-        if len(ids) < 2:
-            notifications.show("Select at least 2 branches to compare.", "warning")
-            await bus.publish("notification")
-            return None
-
-        sources = []
-        for node_id in ids:
-            node = document.nodes.get(node_id)
-            if node is None or node.kind != "chat":
-                notifications.show("Every selected node must be a real chat message to compare.", "warning")
-                await bus.publish("notification")
-                return None
-            sources.append(node)
-
-        branches = [
-            (f"Branch {index + 1}", document.chat_branch_history(node.id))
-            for index, node in enumerate(sources)
-        ]
-        formatted = _format_branches_for_comparison(branches)
-
-        # Positioned below-and-right of the source branches, the same
-        # "offset to the side" convention _generate_note_from_node uses for
-        # a single source - averaged/maxed across all sources here since
-        # there's more than one.
-        avg_x = sum(node.x for node in sources) / len(sources)
-        max_y = max(node.y for node in sources)
-
-        result_holder: dict[str, str] = {}
-
-        async def _on_success(text):
-            if any(node_id not in document.nodes for node_id in ids):
-                # A source was deleted mid-flight - same liveness posture as
-                # _generate_note_from_node's own on_success guard.
-                return
-            note = document.add_note(avg_x + NOTE_AGENT_X_OFFSET, max_y)
-            document.set_note_content(note.id, text)
-            document.set_group_color(note.id, NOTE_AGENT_BODY_COLOR, NOTE_AGENT_HEADER_COLOR)
-            document.mark_branch_comparison_note(note.id, ids)
-            result_holder["node_id"] = note.id
-            await bus.publish("scene")
-
-        def _on_failure(message):
-            # start_branch_comparison already surfaced the notification.
-            pass
-
-        await agent_dispatcher.start_branch_comparison(
-            bus=bus,
-            notifications_state=notifications,
-            source_text=formatted,
-            on_success=_on_success,
-            on_failure=_on_failure,
-        )
-        return result_holder.get("node_id")
-
-    async def synthesize_branches(node_ids, instructions):
-        """ADR-002 Workstream 1 ("Synthesize Branches") - the third
-        sequenced item in that workstream's own "fork -> compare ->
-        synthesize -> status/lifecycle UI" order, following compare_
-        branches above. Same validation contract as that function (2+
-        de-duped ids, every one a real chat node, no auto-selection
-        fallback), plus one more: instructions must be non-blank, since an
-        empty steering prompt would leave the agent nothing to follow.
-
-        Unlike Compare (whose result is a parentless note), Synthesize's
-        result is a real CHAT node continuing the branch tree from the
-        FIRST selected source - a genuine next step in the conversation,
-        not a side annotation - so last_chat_node_id is updated to it
-        exactly like an ordinary send. Every source is still recorded (via
-        item_ids, the same multi-purpose-field reuse Compare's note
-        already established) so full provenance survives even though only
-        one edge can be structural. Provider/model are stamped from
-        composer_document.route() - the same route a plain send would
-        actually use - onto the result node (see SceneNode.provider/model's
-        own comment)."""
-        ids = list(dict.fromkeys(str(i) for i in (node_ids or [])))  # de-dupe, preserve order
-        if len(ids) < 2:
-            notifications.show("Select at least 2 branches to synthesize.", "warning")
-            await bus.publish("notification")
-            return None
-
-        clean_instructions = str(instructions or "").strip()
-        if not clean_instructions:
-            notifications.show("Enter instructions for how to combine the branches.", "warning")
-            await bus.publish("notification")
-            return None
-
-        sources = []
-        for node_id in ids:
-            node = document.nodes.get(node_id)
-            if node is None or node.kind != "chat":
-                notifications.show("Every selected node must be a real chat message to synthesize.", "warning")
-                await bus.publish("notification")
-                return None
-            sources.append(node)
-
-        branches = [
-            (f"Branch {index + 1}", document.chat_branch_history(node.id))
-            for index, node in enumerate(sources)
-        ]
-        formatted = _format_branches_for_comparison(branches)
-
-        parent = sources[0]
-        avg_x = sum(node.x for node in sources) / len(sources)
-        max_y = max(node.y for node in sources)
-        route = composer_document.route()
-
-        result_holder: dict[str, str] = {}
-
-        async def _on_success(text):
-            if any(node_id not in document.nodes for node_id in ids):
-                # A source was deleted mid-flight - same liveness posture as
-                # compare_branches's own on_success guard.
-                return
-            node = document.add_chat_node(
-                avg_x, max_y + MESSAGE_VERTICAL_SPACING, text, False, parent_id=parent.id,
-            )
-            document.mark_branch_synthesis(
-                node.id, ids, clean_instructions, route.get("provider"), route.get("modelLabel"),
-            )
-            document.last_chat_node_id = node.id
-            result_holder["node_id"] = node.id
-            await bus.publish("scene")
-
-        def _on_failure(message):
-            # start_branch_synthesis already surfaced the notification.
-            pass
-
-        await agent_dispatcher.start_branch_synthesis(
-            bus=bus,
-            notifications_state=notifications,
-            source_text=formatted,
-            instructions=clean_instructions,
-            on_success=_on_success,
-            on_failure=_on_failure,
-        )
-        return result_holder.get("node_id")
-
-    async def resize_chart(node_id, width, height):
-        document.resize_chart(node_id, width, height)
-        await publish_scene()
-
-    async def toggle_chart_aspect_lock(node_id):
-        document.toggle_chart_aspect_lock(node_id)
-        await publish_scene()
-
-    # -- R5.3: Gitlink node --------------------------------------------------
-    #
-    # Reuses the existing generic pending_request_id field as the busy/
-    # in-flight marker for every Gitlink action (list repos, load tree,
-    # import, build context, run, apply) - this is exactly that field's
-    # documented purpose, and critically it is what makes the
-    # fingerprint-recheck race-proof: a Run cannot start while an Apply
-    # request_id occupies this node's slot, and vice versa.
-
-    async def fetch_gitlink_repositories(node_id):
-        node = document.nodes.get(node_id)
-        if node is None or node.pending_request_id:
-            notifications.show("Gitlink is busy for this node.", "info")
-            await bus.publish("notification")
-            return []
-        return await agent_dispatcher.fetch_gitlink_repositories(
-            bus=bus, notifications_state=notifications, node=node,
-        )
-
-    async def load_gitlink_repo_tree(node_id, repo, branch):
-        node = document.nodes.get(node_id)
-        if node is None or node.pending_request_id:
-            notifications.show("Gitlink is busy for this node.", "info")
-            await bus.publish("notification")
-            return None
-        result = await agent_dispatcher.load_gitlink_repo_tree(
-            bus=bus, notifications_state=notifications, node=node, repo=repo, branch=branch,
-        )
-        if result is not None:
-            document.store_gitlink_repo_tree(node_id, *result)
-            await publish_scene()
-        return node_id
-
-    async def set_gitlink_local_root(node_id, local_root):
-        document.set_gitlink_local_root(node_id, local_root)
-        await publish_scene()
-
-    async def pick_gitlink_local_root(node_id):
-        # R8a (UI/UX audit POLISH finding #1): the field's own label used to
-        # say "no browse - deferred", dating from before native_dialogs.py
-        # existed. pick_folder is already generic and already used by
-        # Settings' Ollama/Llama.cpp Scan Folder buttons - this just wires
-        # the same primitive to Gitlink's own local-root field instead of
-        # requiring the user to type a path by hand. A cancelled dialog is
-        # a quiet no-op, matching every other pick_folder call site.
-        node = document.nodes.get(node_id)
-        if node is None:
-            return
-        directory = node.gitlink_local_root or os.path.expanduser("~")
-        try:
-            folder = await native_dialogs.pick_folder(directory=directory)
-        except Exception as exc:  # noqa: BLE001 - a local folder path, not a credential
-            notifications.show(f"Could not open the folder picker: {exc}", "error")
-            await bus.publish("notification")
-            return
-        if not folder:
-            return
-        document.set_gitlink_local_root(node_id, folder)
-        await publish_scene()
-
-    async def import_gitlink_snapshot(node_id, repo, branch):
-        node = document.nodes.get(node_id)
-        if node is None or node.pending_request_id:
-            notifications.show("Gitlink is busy for this node.", "info")
-            await bus.publish("notification")
-            return None
-        result = await agent_dispatcher.import_gitlink_snapshot(
-            bus=bus, notifications_state=notifications, node=node, repo=repo, branch=branch,
-            local_root_hint=node.gitlink_local_root, imported_root_hint=node.gitlink_imported_root,
-        )
-        if result is not None:
-            document.store_gitlink_snapshot_root(node_id, *result)
-            await publish_scene()
-        return node_id
-
-    async def build_gitlink_context(node_id, scope_mode, selected_paths):
-        node = document.nodes.get(node_id)
-        if node is None or node.pending_request_id:
-            notifications.show("Gitlink is busy for this node.", "info")
-            await bus.publish("notification")
-            return None
-        result = await agent_dispatcher.build_gitlink_context(
-            bus=bus, notifications_state=notifications, node=node,
-            scope_mode=scope_mode, selected_paths=list(selected_paths),
-        )
-        if result is not None:
-            document.store_gitlink_context(node_id, scope_mode=scope_mode,
-                                            selected_paths=selected_paths, **result)
-            await publish_scene()
-        return node_id
-
-    async def fetch_gitlink_context(node_id):
-        return document.fetch_gitlink_context_xml(node_id)
-
-    async def run_gitlink_change_set(node_id, task_prompt):
-        node_for_check = document.nodes.get(node_id)
-        if node_for_check is not None and node_for_check.pending_request_id:
-            notifications.show("Gitlink is already busy for this node.", "info")
-            await bus.publish("notification")
-            return None
-        # R5.3 post-review FIX 4(b): claim the busy slot with a placeholder
-        # SYNCHRONOUSLY, in the same stretch as the busy pre-check just
-        # above - before document.start_gitlink_run or any await - so a
-        # second concurrent call for this SAME node_id can never pass that
-        # same pre-check during the `await publish_scene()` gap below.
-        # agent_dispatcher.start_gitlink_run (the ONLY caller of this dict
-        # entry for this node_id, invoked just below) recognizes this exact
-        # placeholder and overwrites it with the real request_id, still
-        # synchronously - see that method's own docstring.
-        if node_for_check is not None:
-            node_for_check.pending_request_id = _GITLINK_RUN_CLAIM_PLACEHOLDER
-        try:
-            node = document.start_gitlink_run(node_id, task_prompt)
-        except SceneError:
-            # Node deleted (or wrong-kind) concurrently with the claim above -
-            # the placeholder must not linger on a node this handler is
-            # about to give up on.
-            if node_for_check is not None:
-                node_for_check.pending_request_id = None
-            notifications.show("This node no longer exists.", "warning")
-            await bus.publish("notification")
-            return None
-        await publish_scene()
-
-        def _on_success(proposal_markdown, pending_changes, preview_text, fingerprint, local_root):
-            document.complete_gitlink_run(node_id, proposal_markdown, pending_changes,
-                                           preview_text, fingerprint, local_root)
-
-        def _on_failure(message):
-            document.fail_gitlink_run(node_id, message)
-
-        await agent_dispatcher.start_gitlink_run(
-            bus=bus, notifications_state=notifications, node=node, node_id=node_id,
-            repo=node.gitlink_repo, branch=node.gitlink_branch,
-            scope_mode=node.gitlink_scope_mode, task_prompt=task_prompt,
-            context_xml=node.gitlink_context_xml, context_summary=node.gitlink_context_summary,
-            local_root=node.gitlink_local_root,
-            on_success=_on_success, on_failure=_on_failure,
-        )
-        return node_id
-
-    async def cancel_gitlink_request(request_id):
-        agent_dispatcher.cancel_gitlink(request_id)
-
-    async def apply_gitlink_changes(node_id, fingerprint):
-        node = document.nodes.get(node_id)
-        if node is None:
-            notifications.show("This node no longer exists.", "warning")
-            await bus.publish("notification")
-            return None
-
-        def _on_success(written_files):
-            document.complete_gitlink_apply(node_id, written_files)
-
-        def _on_failure(message):
-            document.fail_gitlink_apply(node_id, message)
-
-        await agent_dispatcher.start_gitlink_apply(
-            bus=bus, notifications_state=notifications, node=node, node_id=node_id,
-            client_fingerprint=fingerprint, local_root=node.gitlink_local_root,
-            on_success=_on_success, on_failure=_on_failure,
-        )
-        return node_id
-
-    bus.register_intent("scene", "fetchGitlinkRepositories", fetch_gitlink_repositories)
-    bus.register_intent("scene", "loadGitlinkRepoTree", load_gitlink_repo_tree)
-    bus.register_intent("scene", "setGitlinkLocalRoot", set_gitlink_local_root)
-    bus.register_intent("scene", "pickGitlinkLocalRoot", pick_gitlink_local_root)
-    bus.register_intent("scene", "importGitlinkSnapshot", import_gitlink_snapshot)
-    bus.register_intent("scene", "buildGitlinkContext", build_gitlink_context)
-    bus.register_intent("scene", "fetchGitlinkContext", fetch_gitlink_context)
-    bus.register_intent("scene", "runGitlinkChangeSet", run_gitlink_change_set)
-    bus.register_intent("scene", "cancelGitlinkRequest", cancel_gitlink_request)
-    # CRITICAL, load-bearing property: applyGitlinkChanges takes ONLY
-    # (node_id, fingerprint) as WS intent arguments - there must be NO
-    # changes/pending_changes parameter anywhere in this signature or the
-    # dispatcher method it calls. This closes the most obvious
-    # content-injection bypass by construction, not by a runtime check: the
-    # only content that ever reaches apply_change_set is server-held,
-    # already-normalized node.gitlink_pending_changes.
-    bus.register_intent("scene", "applyGitlinkChanges", apply_gitlink_changes)
-
-    # -- R5.4: Py-Coder node ---------------------------------------------------
-
-    async def set_pycoder_mode(node_id, mode):
-        document.set_pycoder_mode(node_id, mode)
-        await publish_scene()
-
-    async def run_pycoder(node_id, input_text):
-        # R5.3 post-review FIX 4(b)'s own Run-vs-Run race fix, reused
-        # verbatim for this new kind: claim the busy slot with a shared
-        # placeholder SYNCHRONOUSLY, in the same stretch as the busy
-        # pre-check just above - before document.start_pycoder_run or any
-        # await - so a second concurrent runPyCoder for this SAME node_id
-        # can never pass the same pre-check during the `await
-        # publish_scene()` gap below. Critically, this placeholder stays
-        # claimed for the ENTIRE span from here through generation, through
-        # the human-approval pause, through execution, through analysis - so
-        # a second runPyCoder DURING the pause is refused by this SAME
-        # check, no new logic needed for that case specifically (see the
-        # R5.4 design spec's own section on this).
-        node_for_check = document.nodes.get(node_id)
-        if node_for_check is not None and node_for_check.pending_request_id:
-            notifications.show("Py-Coder is already busy for this node.", "info")
-            await bus.publish("notification")
-            return None
-        if node_for_check is not None:
-            node_for_check.pending_request_id = _CODE_EXEC_RUN_CLAIM_PLACEHOLDER
-        try:
-            node = document.start_pycoder_run(node_id, input_text)
-        except SceneError:
-            if node_for_check is not None:
-                node_for_check.pending_request_id = None
-            notifications.show("This node no longer exists.", "warning")
-            await bus.publish("notification")
-            return None
-        await publish_scene()
-
-        parent_edge = document._branch_parent_edge(node_id)
-        branch_history = document.chat_branch_history(parent_edge.source) if parent_edge else []
-
-        def _on_success(code, output, analysis, last_run_failed):
-            document.complete_pycoder_run(node_id, code, output, analysis, last_run_failed)
-
-        def _on_failure(message):
-            document.fail_pycoder_run(node_id, message)
-
-        await agent_dispatcher.start_pycoder_run(
-            bus=bus, notifications_state=notifications, node=node, node_id=node_id,
-            mode=node.pycoder_mode, prompt=node.pycoder_prompt, code=node.pycoder_code,
-            conversation_history=branch_history,
-            on_success=_on_success, on_failure=_on_failure,
-        )
-        return node_id
-
-    async def cancel_pycoder_request(request_id):
-        agent_dispatcher.cancel_pycoder(request_id)
-
-    bus.register_intent("scene", "setPyCoderMode", set_pycoder_mode)
-    bus.register_intent("scene", "runPyCoder", run_pycoder)
-    bus.register_intent("scene", "cancelPyCoderRequest", cancel_pycoder_request)
-
-    # -- R5.4: Execution Sandbox node -------------------------------------------
-
-    async def set_code_sandbox_requirements(node_id, requirements_text):
-        document.set_code_sandbox_requirements(node_id, requirements_text)
-        await publish_scene()
-
-    async def run_code_sandbox(node_id, input_text):
-        # Same busy-claim-placeholder pattern as run_pycoder above (and
-        # run_gitlink_change_set before it) - see that function's own
-        # comment for the exact race this closes.
-        node_for_check = document.nodes.get(node_id)
-        if node_for_check is not None and node_for_check.pending_request_id:
-            notifications.show("Virtual Environment Runner is already busy for this node.", "info")
-            await bus.publish("notification")
-            return None
-        if node_for_check is not None:
-            node_for_check.pending_request_id = _CODE_EXEC_RUN_CLAIM_PLACEHOLDER
-        try:
-            node = document.start_code_sandbox_run(node_id, input_text)
-        except SceneError:
-            if node_for_check is not None:
-                node_for_check.pending_request_id = None
-            notifications.show("This node no longer exists.", "warning")
-            await bus.publish("notification")
-            return None
-        await publish_scene()
-
-        parent_edge = document._branch_parent_edge(node_id)
-        branch_history = document.chat_branch_history(parent_edge.source) if parent_edge else []
-
-        def _on_success(code, output, analysis):
-            document.complete_code_sandbox_run(node_id, code, output, analysis)
-
-        def _on_failure(message):
-            document.fail_code_sandbox_run(node_id, message)
-
-        await agent_dispatcher.start_code_sandbox_run(
-            bus=bus, notifications_state=notifications, node=node, node_id=node_id,
-            sandbox_id=node.code_sandbox_sandbox_id,
-            prompt=node.code_sandbox_prompt, existing_code=node.code_sandbox_code,
-            requirements_manifest=node.code_sandbox_requirements,
-            conversation_history=branch_history,
-            on_success=_on_success, on_failure=_on_failure,
-        )
-        return node_id
-
-    async def cancel_code_sandbox_request(request_id):
-        agent_dispatcher.cancel_code_sandbox(request_id)
-
-    bus.register_intent("scene", "setCodeSandboxRequirements", set_code_sandbox_requirements)
-    bus.register_intent("scene", "runCodeSandbox", run_code_sandbox)
-    bus.register_intent("scene", "cancelCodeSandboxRequest", cancel_code_sandbox_request)
-
-    # -- R5.4: shared approve/deny - one request_id namespace across both kinds
-
-    async def approve_code_execution(request_id):
-        agent_dispatcher.approve_code_execution(request_id)
-
-    async def deny_code_execution(request_id):
-        agent_dispatcher.deny_code_execution(request_id)
-
-    bus.register_intent("scene", "approveCodeExecution", approve_code_execution)
-    bus.register_intent("scene", "denyCodeExecution", deny_code_execution)
+    register_web_research_intents(bus, document, notifications, agent_dispatcher)
+    register_artifact_intents(bus, document, notifications, agent_dispatcher)
+    register_chart_intents(bus, document, notifications, agent_dispatcher)
+    register_branches_intents(bus, document, notifications, agent_dispatcher, composer_document)
+    register_gitlink_intents(bus, document, notifications, agent_dispatcher)
+    register_pycoder_intents(bus, document, notifications, agent_dispatcher)
+    register_code_sandbox_intents(bus, document, notifications, agent_dispatcher)
 
     # -- R6.1: Notes/Frames/Containers -----------------------------------------
 
@@ -1136,29 +432,6 @@ def register_canvas(
         document.set_view_state(zoom_factor, scroll_x, scroll_y)
         await publish_scene()
 
-    # R5.1: Web Research node run/cancel - node creation itself lives in
-    # backend/plugins.py's executePlugin (the "Web Research" branch), not
-    # here; these two intents drive an EXISTING web_research-kind node.
-    bus.register_intent("scene", "runWebResearch", run_web_research)
-    bus.register_intent("scene", "cancelWebResearchRequest", cancel_web_research_request)
-    # R5.2: Artifact/Drafter Send/cancel - node creation itself lives in
-    # backend/plugins.py's executePlugin (the "Artifact / Drafter" branch),
-    # not here; these two intents drive an EXISTING artifact-kind node, same
-    # posture as Web Research's own runWebResearch/cancelWebResearchRequest
-    # pair above.
-    bus.register_intent("scene", "sendArtifactMessage", send_artifact_message)
-    bus.register_intent("scene", "cancelArtifactRequest", cancel_artifact_request)
-    # R6.2: Chart - a single combined create+generate action, unlike every
-    # node-creation flow above - see generate_chart's own docstring.
-    bus.register_intent("scene", "generateChart", generate_chart)
-    # R8a: the two note agents restored from the deleted Qt app - see
-    # graphlink_note_agent.py's own docstring for why they were dead stubs.
-    bus.register_intent("scene", "generateKeyTakeaway", generate_key_takeaway)
-    bus.register_intent("scene", "generateExplainerNote", generate_explainer_note)
-    bus.register_intent("scene", "compareBranches", compare_branches)
-    bus.register_intent("scene", "synthesizeBranches", synthesize_branches)
-    bus.register_intent("scene", "resizeChart", resize_chart)
-    bus.register_intent("scene", "toggleChartAspectLock", toggle_chart_aspect_lock)
     bus.register_intent("scene", "addPin", add_pin)
     bus.register_intent("scene", "movePin", move_pin)
     bus.register_intent("scene", "removePin", remove_pin)
